+++ debian/.gitignore
@@ -23,7 +23,7 @@
 /sprout-bgcf-dbg/
 /sprout-mmtel-as/
 /sprout-mmtel-as-dbg/
-/gemini/
-/gemini-dbg/
-/memento-sip/
-/memento-sip-dbg/
+/gemini-as/
+/gemini-as-dbg/
+/memento-as/
+/memento-as-dbg/
