+++ debian/gemini-as.install
@@ -1,0 +1,1 @@
+build/bin/gemini-as.so usr/share/clearwater/sprout/plugins
