--- debian/gemini.install
@@ -1,1 +1,0 @@
-build/bin/gemini.so usr/share/clearwater/sprout/plugins
