+++ debian/memento-as.install
@@ -1,0 +1,1 @@
+build/bin/memento-as.so usr/share/clearwater/sprout/plugins
