+++ debian/memento-as.postrm
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-# @file memento-sip.postinst
+# @file memento-as.postrm
 #
 # Project Clearwater - IMS in the Cloud
 # Copyright (C) 2014  Metaswitch Networks Ltd
@@ -39,34 +39,7 @@
 
 set -e
 
-# summary of how this script can be called:
-#        * <postinst> `configure' <most-recently-configured-version>
-#        * <old-postinst> `abort-upgrade' <new version>
-#        * <conflictor's-postinst> `abort-remove' `in-favour' <package>
-#          <new-version>
-#        * <postinst> `abort-remove'
-#        * <deconfigured's-postinst> `abort-deconfigure' `in-favour'
-#          <failed-install-package> <version> `removing'
-#          <conflicting-package> <version>
-# for details, see http://www.debian.org/doc/debian-policy/ or
-# the debian-policy package
-
-case "$1" in
-    configure)
-        # Restart sprout to force reload of new plug-in.  Always do this by
-        # terminating sprout so monit will restart it more-or-less immediately.
-        # (monit restart seems to have significant lag.)
-        # Don't fail if it's already stopped.
-        service sprout stop || /bin/true
-    ;;
-
-    abort-upgrade|abort-remove|abort-deconfigure)
-    ;;
-
-    *)
-        echo "postinst called with unknown argument \`$1'" >&2
-        exit 1
-    ;;
-esac
+# Restart sprout so it unloads the plug-in that has been removed.
+sudo service sprout stop || /bin/true
 
 exit 0
