+++ debian/memento-as.prerm
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-# @file gemini.prerm
+# @file memento-as.prerm
 #
 # Project Clearwater - IMS in the Cloud
 # Copyright (C) 2014  Metaswitch Networks Ltd
