--- debian/memento-sip.install
@@ -1,1 +1,0 @@
-build/bin/memento-sip.so usr/share/clearwater/sprout/plugins
