+++ debian/rules
@@ -41,7 +41,7 @@ override_dh_strip:
 	dh_strip -psprout-icscf --dbg-package=sprout-icscf-dbg
 	dh_strip -psprout-bgcf --dbg-package=sprout-bgcf-dbg
 	dh_strip -psprout-mmtel-as --dbg-package=sprout-mmtel-as-dbg
-	dh_strip -pgemini --dbg-package=gemini-dbg
-	dh_strip -pmemento-sip --dbg-package=memento-sip-dbg
+	dh_strip -pgemini-as --dbg-package=gemini-as-dbg
+	dh_strip -pmemento-as --dbg-package=memento-as-dbg
 	dh_strip -pbono --dbg-package=bono-dbg
 	dh_strip -pclearwater-sip-stress --dbg-package=clearwater-sip-stress-dbg
