+++ sprout/Makefile
@@ -7,8 +7,8 @@ all:
 	make -f sprout_icscf.mk
 	make -f sprout_bgcf.mk
 	make -f sprout_mmtel_as.mk
-	make -f gemini.mk
-	make -f memento-sip.mk
+	make -f gemini_as.mk
+	make -f memento_as.mk
 
 .PHONY: test run_test coverage coverage-check coverage_raw debug vg vg-check vg_raw
 test run_test coverage coverage-check coverage_raw debug vg vg-check vg_raw:
@@ -21,6 +21,6 @@ clean:
 	make -f sprout_icscf.mk clean
 	make -f sprout_bgcf.mk clean
 	make -f sprout_mmtel_as.mk clean
-	make -f gemini.mk -C clean
-	make -f memento-sip.mk clean
+	make -f gemini_as.mk clean
+	make -f memento_as.mk clean
 	make -f sprout_test.mk clean
